+++ @apicize/toolkit/src/index.ts
@@ -11,6 +11,7 @@ export { CertificateFileType } from './models/workspace/editable-certificate'
 export { NavigationControl } from './controls/navigation/navigation'
 export { HelpPanel } from './controls/help'
 export { MainPanel } from './controls/main-panel'
+export { HelpContents } from './models/help-contents'
 
 export * from './theme'
 export * from './services/base64'