+++ @apicize/toolkit/src/models/help-contents.ts
@@ -1,0 +1,4 @@
+/**
+ * Hierarchical structure to store Help Contents in
+ */
+export type HelpContents = { [name: string]: (string | HelpContents) }